Create cars controller
